Engagement at BluePearl
This course was developed along with BluePearl's Talent Development Team for the purpose of explaining the purpose and method for implementing the yearly Engagement Survey. I used PowerPoint and Microsoft Office to rapidly design this course along with the Talent Development Team. I then used Articulate Storyline to develop the course in 3 iterrations.